+++ web-next/wasm/src/repository/local_resource.rs
@@ -46,7 +46,8 @@ impl IdbId for IdbIdWrapper<LocalResourceId> {
 
         Ok(IdbIdWrapper(LocalResourceId {
             path: json_array.first().and_then(|it| serde_json::from_value(it.clone()).ok()),
-            order_id: json_array.get(2).and_then(|it| it.as_str().and_then(|it| it.parse().ok()))
+            order_id: json_array.get(2).and_then(|it| it.as_str().and_then(|it| it.parse().ok())),
+            shelf_id: None
         }))
     }
 }
@@ -197,10 +198,11 @@ impl LocalResourceRepository for LocalResourceRepositoryImpl {
         Ok(reader.entry().await?.size)
     }
 
-    async fn remove(&self, path: LocalResourcePath) -> Result<Vec<LocalResource>, PersistenceError> {
+    async fn remove(&self, path: LocalResourcePath, shelf_id: u64) -> Result<Vec<LocalResource>, PersistenceError> {
         let from_id = LocalResourceId {
             path: Some(path),
-            order_id: None
+            order_id: None,
+            shelf_id: Some(shelf_id)
         };
 
         let items = IdbRepository::<LocalResource, IdbIdWrapper<LocalResourceId>>::find_all(
@@ -212,6 +214,9 @@ impl LocalResourceRepository for LocalResourceRepositoryImpl {
 
         let mut removed_items = vec![];
         for item in items.iter() {
+            if item.shelf_id != shelf_id {
+                continue;
+            }
             let id: LocalResourceId = Table::<LocalResourceId>::id(item);
             let removed = Repository::<LocalResource, LocalResourceId>::delete_one(self, &id).await;
             if let Ok(item) = removed {
